# uzum_api_gateway/api/docs/swagger.yaml
definitions:
...
  models.ResponseError:
    properties:
      error: {}
    type: object
...
  user_service.CreateSeller:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      shop_id:
        type: string
    type: object
...
  user_service.CreateUs:
    properties:
      gmail:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
...
  user_service.Empty:
    type: object
  user_service.Empty2:
    type: object
...
  user_service.GetListSellerResponse:
    properties:
      count:
        type: integer
      sellers:
        items:
          $ref: '#/definitions/user_service.Seller'
        type: array
    type: object
...
  user_service.GetListUsResponse:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/user_service.Us'
        type: array
    type: object
  user_service.Seller:
    properties:
      created_at:
        type: string
      deleted_at:
        type: integer
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      shop_id:
        type: string
      updated_at:
        type: string
    type: object
...
  user_service.UpdateSeller:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      shop_id:
        type: string
    type: object
...
  user_service.UpdateUs:
    properties:
      gmail:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  user_service.Us:
    properties:
      created_at:
        type: string
      deleted_at:
        type: integer
      gmail:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /CreateSeller:
    post:
      consumes:
      - application/json
      description: API for creating seller
      parameters:
      - description: seller
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateSeller'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.CreateSeller'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create seller
      tags:
      - seller
...
  /CreateUser:
    post:
      consumes:
      - application/json
      description: API for creating user
      parameters:
      - description: user
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateUs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.CreateUs'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create user
      tags:
      - user
  /DeleteSeller/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting a seller by ID
      parameters:
      - description: seller ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Empty2'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a seller by ID
      tags:
      - seller
...
  /DeleteUser/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting a user by ID
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a user by ID
      tags:
      - user
  /GetByIdSeller/{id}:
    get:
      consumes:
      - application/json
      description: API for getting a single seller by ID
      parameters:
      - description: seller ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Seller'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a single seller by ID
      tags:
      - seller
...
  /GetByIdUser/{id}:
    get:
      consumes:
      - application/json
      description: API for getting a single user by ID
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Us'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a single user by ID
      tags:
      - user
...
  /GetListSeller:
    get:
      consumes:
      - application/json
      description: API for getting list seller
      parameters:
      - description: sellers
        in: query
        name: seller
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetListSellerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get List seller
      tags:
      - seller
...
  /GetListUser:
    get:
      consumes:
      - application/json
      description: API for getting list user
      parameters:
      - description: system_users
        in: query
        name: system_user
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetListUsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get List user
      tags:
      - user
...
  /UpdateSeller/{id}:
    put:
      consumes:
      - application/json
      description: API for updating a seller by ID
      parameters:
      - description: seller ID
        in: path
        name: id
        required: true
        type: string
      - description: seller
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateSeller'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.UpdateSeller'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a seller by ID
      tags:
      - seller
...
  /UpdateUser/{id}:
    put:
      consumes:
      - application/json
      description: API for updating a user by ID
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: string
      - description: user
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateUs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Us'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a user by ID
      tags:
      - user
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"